Mudança de tipos

+++ New Unity Project 1/Assets/CSs/SliderController.cs
@@ -7,7 +7,7 @@ namespace Assets.CSs
     class SliderController : MonoBehaviour
     {
          
-        void SettingPlayers(Single value)
+        void SettingPlayers(int value)
         {
             int qtIA = 1;
             int qtPlayer = 2;
